requirements.txt
numpy
pandas
matplotlib
streamlit
python-pptx
